Move reducer for position to forest/screen.py
This reduces screen actions so makes more sense here.
